# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrightRoots Adventure API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileAgeGroup = typeof ProfileAgeGroup[keyof typeof ProfileAgeGroup];


export const ProfileAgeGroup = {
  '2-3': '2-3',
  '4-5': '4-5',
  '6-7': '6-7',
} as const;

export type ProfileProgress = {[key: string]: number};

export type ProfileAccuracy = {[key: string]: number};

export interface Profile {
  id: number;
  name: string;
  ageGroup: ProfileAgeGroup;
  coins: number;
  streak: number;
  progress?: ProfileProgress;
  accuracy?: ProfileAccuracy;
  createdAt: string;
}

export type ProfileInputAgeGroup = typeof ProfileInputAgeGroup[keyof typeof ProfileInputAgeGroup];


export const ProfileInputAgeGroup = {
  '2-3': '2-3',
  '4-5': '4-5',
  '6-7': '6-7',
} as const;

export interface ProfileInput {
  /** @minLength 1 */
  name: string;
  ageGroup: ProfileInputAgeGroup;
}

export interface ProgressInput {
  subject?: string;
  correct?: boolean;
  coins?: number;
  streak?: number;
}

export interface SubjectStat {
  subject: string;
  correct: number;
  total: number;
  accuracy: number;
}

export interface ProfileStats {
  profileId: number;
  /** @nullable */
  strongestSubject: string | null;
  /** @nullable */
  weakestSubject: string | null;
  totalLessons: number;
  totalCoins: number;
  overallAccuracy: number;
  subjectBreakdown?: SubjectStat[];
}

export interface PinInput {
  pin: string;
}

export interface PinToken {
  token: string;
}