# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrightRoots Adventure API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: profiles
    description: Child profile operations
  - name: progress
    description: Progress tracking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: List all child profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a child profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{id}/progress:
    put:
      operationId: updateProgress
      tags: [progress]
      summary: Update progress for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /auth/pin:
    post:
      operationId: verifyParentPin
      tags: [auth]
      summary: Verify parent PIN and receive a signed JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "200":
          description: Signed token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinToken"
        "401":
          description: Incorrect PIN
        "503":
          description: Server misconfiguration
  /profiles/{id}/reset:
    put:
      operationId: resetProfile
      tags: [progress]
      summary: Reset all progress for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
  /profiles/{id}/stats:
    get:
      operationId: getProfileStats
      tags: [progress]
      summary: Get detailed stats for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Profile:
      type: object
      required: [id, name, ageGroup, coins, streak, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        ageGroup:
          type: string
          enum: ["2-3", "4-5", "6-7"]
        coins:
          type: integer
        streak:
          type: integer
        progress:
          type: object
          additionalProperties:
            type: integer
        accuracy:
          type: object
          additionalProperties:
            type: number
        createdAt:
          type: string
    ProfileInput:
      type: object
      required: [name, ageGroup]
      properties:
        name:
          type: string
          minLength: 1
        ageGroup:
          type: string
          enum: ["2-3", "4-5", "6-7"]
    ProgressInput:
      type: object
      properties:
        subject:
          type: string
        correct:
          type: boolean
        coins:
          type: integer
        streak:
          type: integer
    ProfileStats:
      type: object
      required: [profileId, strongestSubject, weakestSubject, totalLessons, totalCoins, overallAccuracy]
      properties:
        profileId:
          type: integer
        strongestSubject:
          type: ["string", "null"]
        weakestSubject:
          type: ["string", "null"]
        totalLessons:
          type: integer
        totalCoins:
          type: integer
        overallAccuracy:
          type: number
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectStat"
    SubjectStat:
      type: object
      required: [subject, correct, total, accuracy]
      properties:
        subject:
          type: string
        correct:
          type: integer
        total:
          type: integer
        accuracy:
          type: number
    PinInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
    PinToken:
      type: object
      required: [token]
      properties:
        token:
          type: string